requirements.txt
# Azure AI Projects & Agents SDK
azure-ai-projects>=1.0.0b10
azure-ai-agents>=1.0.0
azure-identity>=1.15.0

# Streamlit for UI
streamlit>=1.31.0

# HTTP requests (for MCP server calls)
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0

# Type hints
typing-extensions>=4.9.0

# OpenTelemetry Tracing
opentelemetry-sdk>=1.25.0
azure-monitor-opentelemetry>=1.6.0
azure-core-tracing-opentelemetry>=1.0.0b11
opentelemetry-instrumentation-openai-v2>=2.0.0
